fix tslint issues
