+++ test/entry.ts
@@ -1,5 +1,5 @@
 /// <reference types="mocha" />
 
-import "bluebird/js/browser/bluebird";
-import "es6-symbol/implement";
-import "es6-map/implement";
+import "bluebird";
+import "es6-symbol";
+import "es6-map";
