+++ test/mock/MockEvent.ts
@@ -1,9 +1,5 @@
-export class Sprite {}
-
 export class MockEvent {
     constructor(public type: string) {
         this.type = type;
     }
 }
-
-export class MockMouseEvent extends MockEvent {}
