+++ test/mock/MockMouseEvent.ts
@@ -1,0 +1,3 @@
+import { MockEvent } from "./MockEvent";
+
+export class MockMouseEvent extends MockEvent {}
