+++ test/mock/Sprite.ts
@@ -1,0 +1,1 @@
+export class Sprite {}
