+++ test/org/osflash/signals/DeluxeSignalWithGenericEventTest.test.ts
@@ -5,7 +5,7 @@ import { assert } from "chai";
 import { AsyncUtil } from "../../../util/AsyncUtil";
 import { DeluxeSignal } from "../../../../src/org/osflash/signals/DeluxeSignal";
 import { GenericEvent } from "../../../../src/org/osflash/signals/events/GenericEvent";
-import { Sprite } from "../../../mock/mock";
+import { Sprite } from "../../../mock/Sprite";
 import { IEvent } from "../../../../src/org/osflash/signals/events/IEvent";
 import { newEmptyHandler, failIfCalled } from "../../../util/TestBase";
