+++ test/org/osflash/signals/SignalTestBase.test.ts
@@ -5,7 +5,7 @@ import { assert } from "chai";
 import { ISignal } from "../../../../src/org/osflash/signals/ISignal";
 import { ISlot } from "../../../../src/org/osflash/signals/ISlot";
 import { Signal } from "../../../../src/org/osflash/signals/Signal";
-import { MockEvent } from "../../../mock/mock";
+import { MockEvent } from "../../../mock/MockEvent";
 import { AsyncUtil } from "../../../util/AsyncUtil";
 import { newEmptyHandler, failIfCalled } from "../../../util/TestBase";
