+++ test/org/osflash/signals/support/Base.ts
@@ -1,0 +1,1 @@
+export class Base {}
