+++ test/org/osflash/signals/support/Inherited1.ts
@@ -1,0 +1,3 @@
+import { Base } from "./Base";
+
+export class Inherited1 extends Base {}
