+++ test/org/osflash/signals/support/Inherited2.ts
@@ -1,0 +1,3 @@
+import { Inherited1 } from "./Inherited1";
+
+export class Inherited2 extends Inherited1 {}
